+++ common/chain_test.go
@@ -1,8 +1,13 @@
 package common
 
 import (
+	"encoding/hex"
 	"testing"
 
+	"github.com/btcsuite/btcd/btcec"
+	"github.com/btcsuite/btcd/chaincfg"
+	"github.com/btcsuite/btcutil"
+	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
 )
 
@@ -15,7 +20,7 @@ func TestChain_EncodeAddress(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "should error if b is not a valid address on the network",
+			name: "should error if b is not a valid address on the bitcoin network",
 			chain: Chain{
 				ChainName: ChainName_btc_testnet,
 				ChainId:   18332,
@@ -34,6 +39,36 @@ func TestChain_EncodeAddress(t *testing.T) {
 			want:    "bc1qk0cc73p8m7hswn8y2q080xa4e5pxapnqgp7h9c",
 			wantErr: false,
 		},
+		{
+			name: "should error if b is not a valid address on the evm network",
+			chain: Chain{
+				ChainName: ChainName_goerli_testnet,
+				ChainId:   5,
+			},
+			b:       ethcommon.Hex2Bytes("0x321"),
+			want:    "",
+			wantErr: true,
+		},
+		{
+			name: "should pass if b is a valid address on the evm network",
+			chain: Chain{
+				ChainName: ChainName_goerli_testnet,
+				ChainId:   5,
+			},
+			b:       []byte("0x321"),
+			want:    "0x0000000000000000000000000000003078333231",
+			wantErr: false,
+		},
+		{
+			name: "should error if chain not supported",
+			chain: Chain{
+				ChainName: 999,
+				ChainId:   999,
+			},
+			b:       ethcommon.Hex2Bytes("0x321"),
+			want:    "",
+			wantErr: true,
+		},
 	}
 
 	for _, tc := range tests {
@@ -49,10 +84,298 @@ func TestChain_EncodeAddress(t *testing.T) {
 	}
 }
 
+func TestChain_DecodeAddress(t *testing.T) {
+	tests := []struct {
+		name    string
+		chain   Chain
+		b       string
+		want    []byte
+		wantErr bool
+	}{
+		{
+			name: "should decode on btc chain",
+			chain: Chain{
+				ChainName: ChainName_btc_testnet,
+				ChainId:   18332,
+			},
+			want:    []byte("bc1qk0cc73p8m7hswn8y2q080xa4e5pxapnqgp7h9c"),
+			b:       "bc1qk0cc73p8m7hswn8y2q080xa4e5pxapnqgp7h9c",
+			wantErr: false,
+		},
+		{
+			name: "should decode on evm chain",
+			chain: Chain{
+				ChainName: ChainName_goerli_testnet,
+				ChainId:   5,
+			},
+			want:    ethcommon.HexToAddress("0x321").Bytes(),
+			b:       "0x321",
+			wantErr: false,
+		},
+		{
+			name: "should error if chain not supported",
+			chain: Chain{
+				ChainName: 999,
+				ChainId:   999,
+			},
+			want:    ethcommon.Hex2Bytes("0x321"),
+			b:       "",
+			wantErr: true,
+		},
+	}
+
+	for _, tc := range tests {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			s, err := tc.chain.DecodeAddress(tc.b)
+			if tc.wantErr {
+				require.Error(t, err)
+				return
+			}
+			require.Equal(t, tc.want, s)
+		})
+	}
+}
+
 func TestChain_InChainList(t *testing.T) {
 	require.True(t, ZetaChainMainnet().InChainList(ZetaChainList()))
 	require.True(t, ZetaMocknetChain().InChainList(ZetaChainList()))
 	require.True(t, ZetaPrivnetChain().InChainList(ZetaChainList()))
 	require.True(t, ZetaTestnetChain().InChainList(ZetaChainList()))
 	require.False(t, EthChain().InChainList(ZetaChainList()))
+}
+
+func TestIsZetaChain(t *testing.T) {
+	tests := []struct {
+		name    string
+		chainID int64
+		want    bool
+	}{
+		{"Zeta Mainnet", ZetaChainMainnet().ChainId, true},
+		{"Zeta Testnet", ZetaTestnetChain().ChainId, true},
+		{"Zeta Mocknet", ZetaMocknetChain().ChainId, true},
+		{"Zeta Privnet", ZetaPrivnetChain().ChainId, true},
+		{"Non-Zeta", EthChain().ChainId, false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require.Equal(t, tt.want, IsZetaChain(tt.chainID))
+		})
+	}
+}
+
+func TestIsEVMChain(t *testing.T) {
+	tests := []struct {
+		name    string
+		chainID int64
+		want    bool
+	}{
+		{"Ethereum Mainnet", EthChain().ChainId, true},
+		{"Goerli Testnet", GoerliChain().ChainId, true},
+		{"Sepolia Testnet", SepoliaChain().ChainId, true},
+		{"Non-EVM", BtcMainnetChain().ChainId, false},
+		{"Zeta Mainnet", ZetaChainMainnet().ChainId, false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require.Equal(t, tt.want, IsEVMChain(tt.chainID))
+		})
+	}
+}
+
+func TestIsHeaderSupportedEVMChain(t *testing.T) {
+	tests := []struct {
+		name    string
+		chainID int64
+		want    bool
+	}{
+		{"Ethereum Mainnet", EthChain().ChainId, true},
+		{"Goerli Testnet", GoerliChain().ChainId, true},
+		{"Goerli Localnet", GoerliLocalnetChain().ChainId, true},
+		{"Sepolia Testnet", SepoliaChain().ChainId, true},
+		{"BSC Testnet", BscTestnetChain().ChainId, true},
+		{"BSC Mainnet", BscMainnetChain().ChainId, true},
+		{"Non-EVM", BtcMainnetChain().ChainId, false},
+		{"Zeta Mainnet", ZetaChainMainnet().ChainId, false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require.Equal(t, tt.want, IsHeaderSupportedEvmChain(tt.chainID))
+		})
+	}
+}
+
+func TestSupportMerkleProof(t *testing.T) {
+	tests := []struct {
+		name  string
+		chain Chain
+		want  bool
+	}{
+		{"Ethereum Mainnet", EthChain(), true},
+		{"BSC Testnet", BscTestnetChain(), true},
+		{"BSC Mainnet", BscMainnetChain(), true},
+		{"Non-EVM", BtcMainnetChain(), true},
+		{"Zeta Mainnet", ZetaChainMainnet(), false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require.Equal(t, tt.want, tt.chain.SupportMerkleProof())
+		})
+	}
+}
+
+func TestIsBitcoinChain(t *testing.T) {
+	tests := []struct {
+		name    string
+		chainID int64
+		want    bool
+	}{
+		{"Bitcoin Mainnet", BtcMainnetChain().ChainId, true},
+		{"Bitcoin Testnet", BtcTestNetChain().ChainId, true},
+		{"Bitcoin Regtest", BtcRegtestChain().ChainId, true},
+		{"Non-Bitcoin", EthChain().ChainId, false},
+		{"Zeta Mainnet", ZetaChainMainnet().ChainId, false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require.Equal(t, tt.want, IsBitcoinChain(tt.chainID))
+		})
+	}
+}
+
+func TestIsEthereumChain(t *testing.T) {
+	tests := []struct {
+		name    string
+		chainID int64
+		want    bool
+	}{
+		{"Ethereum Mainnet", EthChain().ChainId, true},
+		{"Goerli Testnet", GoerliChain().ChainId, true},
+		{"Sepolia Testnet", SepoliaChain().ChainId, true},
+		{"Non-Ethereum", BtcMainnetChain().ChainId, false},
+		{"Zeta Mainnet", ZetaChainMainnet().ChainId, false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require.Equal(t, tt.want, IsEthereumChain(tt.chainID))
+		})
+	}
+}
+
+func TestChain_IsExternalChain(t *testing.T) {
+	require.False(t, ZetaChainMainnet().IsExternalChain())
+	require.True(t, EthChain().IsExternalChain())
+}
+
+func TestChain_IsZetaChain(t *testing.T) {
+	require.True(t, ZetaChainMainnet().IsZetaChain())
+	require.False(t, EthChain().IsZetaChain())
+}
+
+func TestChain_IsEmpty(t *testing.T) {
+	require.True(t, Chain{}.IsEmpty())
+	require.False(t, ZetaChainMainnet().IsEmpty())
+}
+
+func TestChain_WitnessProgram(t *testing.T) {
+	// Ordinarily the private key would come from whatever storage mechanism
+	// is being used, but for this example just hard code it.
+	privKeyBytes, err := hex.DecodeString("22a47fa09a223f2aa079edf85a7c2" +
+		"d4f8720ee63e502ee2869afab7de234b80c")
+	require.NoError(t, err)
+
+	t.Run("should return btc address", func(t *testing.T) {
+		_, pubKey := btcec.PrivKeyFromBytes(btcec.S256(), privKeyBytes)
+		pubKeyHash := btcutil.Hash160(pubKey.SerializeCompressed())
+		addr, err := btcutil.NewAddressWitnessPubKeyHash(pubKeyHash, &chaincfg.RegressionNetParams)
+		require.NoError(t, err)
+
+		chain := BtcTestNetChain()
+		_, err = chain.BTCAddressFromWitnessProgram(addr.WitnessProgram())
+		require.NoError(t, err)
+	})
+
+	t.Run("should fail for wrong chain id", func(t *testing.T) {
+		_, pubKey := btcec.PrivKeyFromBytes(btcec.S256(), privKeyBytes)
+		pubKeyHash := btcutil.Hash160(pubKey.SerializeCompressed())
+		addr, err := btcutil.NewAddressWitnessPubKeyHash(pubKeyHash, &chaincfg.RegressionNetParams)
+		require.NoError(t, err)
+
+		chain := GoerliChain()
+		_, err = chain.BTCAddressFromWitnessProgram(addr.WitnessProgram())
+		require.Error(t, err)
+	})
+
+	t.Run("should fail for wrong witness program", func(t *testing.T) {
+		_, pubKey := btcec.PrivKeyFromBytes(btcec.S256(), privKeyBytes)
+		pubKeyHash := btcutil.Hash160(pubKey.SerializeCompressed())
+		addr, err := btcutil.NewAddressWitnessPubKeyHash(pubKeyHash, &chaincfg.RegressionNetParams)
+		require.NoError(t, err)
+
+		chain := BtcTestNetChain()
+		_, err = chain.BTCAddressFromWitnessProgram(addr.WitnessProgram()[0:19])
+		require.Error(t, err)
+	})
+}
+
+func TestChains_Has(t *testing.T) {
+	chains := Chains{ZetaChainMainnet(), ZetaTestnetChain()}
+	require.True(t, chains.Has(ZetaChainMainnet()))
+	require.False(t, chains.Has(EthChain()))
+}
+
+func TestChains_Distinct(t *testing.T) {
+	chains := Chains{ZetaChainMainnet(), ZetaChainMainnet(), ZetaTestnetChain()}
+	distinctChains := chains.Distinct()
+	require.Len(t, distinctChains, 2)
+}
+
+func TestChains_Strings(t *testing.T) {
+	chains := Chains{ZetaChainMainnet(), ZetaTestnetChain()}
+	strings := chains.Strings()
+	expected := []string{chains[0].String(), chains[1].String()}
+	require.Equal(t, expected, strings)
+}
+
+func TestGetChainFromChainID(t *testing.T) {
+	chain := GetChainFromChainID(ZetaChainMainnet().ChainId)
+	require.Equal(t, ZetaChainMainnet(), *chain)
+	require.Nil(t, GetChainFromChainID(9999))
+}
+
+func TestGetBTCChainParams(t *testing.T) {
+	params, err := GetBTCChainParams(BtcMainnetChain().ChainId)
+	require.NoError(t, err)
+	require.Equal(t, &chaincfg.MainNetParams, params)
+
+	_, err = GetBTCChainParams(9999)
+	require.Error(t, err)
+}
+
+func TestGetBTCChainIDFromChainParams(t *testing.T) {
+	chainID, err := GetBTCChainIDFromChainParams(&chaincfg.MainNetParams)
+	require.NoError(t, err)
+	require.Equal(t, int64(8332), chainID)
+
+	chainID, err = GetBTCChainIDFromChainParams(&chaincfg.RegressionNetParams)
+	require.NoError(t, err)
+	require.Equal(t, int64(18444), chainID)
+
+	chainID, err = GetBTCChainIDFromChainParams(&chaincfg.TestNet3Params)
+	require.NoError(t, err)
+	require.Equal(t, int64(18332), chainID)
+
+	_, err = GetBTCChainIDFromChainParams(&chaincfg.Params{Name: "unknown"})
+	require.Error(t, err)
+}
+
+func TestChainIDInChainList(t *testing.T) {
+	require.True(t, ChainIDInChainList(ZetaChainMainnet().ChainId, ZetaChainList()))
+	require.False(t, ChainIDInChainList(EthChain().ChainId, ZetaChainList()))
 }